fix: :zap: adjust user structure - fix calls and unit tests
