+++ src/users/application/create-user.service.ts
@@ -1,9 +1,9 @@
 import { Inject, Injectable } from '@nestjs/common';
 import { USER_REPOSITORY } from '../users.constants';
-import { UserRepository } from '../domain/UserRepository';
+import { UserRepository } from '../domain/user-repository';
 import { User } from '../domain/User';
 import { DomainEventPublisher } from '../../shared/domain/DomainEventPublisher';
-import { UserCreatedEvent } from '../domain/UserCreatedEvent';
+import { UserCreatedEvent } from '../domain/user-created-event';
 
 @Injectable()
 export class CreateUserService {
